Fix: 회원 탈퇴시 파일 값 존재 여부 확인 후 S3에 삭제 요청 전송 #196

+++ back-end/reacton/src/main/java/com/softeer/reacton/domain/professor/ProfessorService.java
@@ -75,8 +75,9 @@ public class ProfessorService {
     public void delete(String oauthId) {
         Professor professor = professorRepository.findByOauthId(oauthId)
                 .orElseThrow(() -> new BaseException(ProfessorErrorCode.PROFESSOR_NOT_FOUND));
-        s3Service.deleteFile(professor.getProfileImageS3Key());
-
+        if (isFileExists(professor)) {
+            s3Service.deleteFile(professor.getProfileImageS3Key());
+        }
         courseRepository.findByProfessor(professor).forEach(course -> {
             scheduleRepository.deleteAllByCourse((Course) course);
             questionRepository.deleteAllByCourse((Course) course);
